Experiment with removing css grid
Removing grid would be beneficial for a gatsby-theme, as it would allow the components
to be cherrypicked for any project. With grid, this is impossible since they are
tied to each other, and the parent grid container needs to always be imported too.

However, there doesn't seem to be a straightforward way to automatically resize
the header and content components when the sidebar is activated/deactivated.
Since this would require Javascript, it would tie these components' state together
anyway, obsoleting the entire point of removing css grid.

Definitely revisit this when a simple resizing solution presents itself!