.header .leftActive/.rightActive bugfix
Whether the bug is actually fixed or obfuscated is up for debate;
this bug occurs because .left/rightActive should only be true
when active in a non-small viewport. However, the !this.isSmallViewport
on l. 157-158 is not correctly passed on because it is not in state,
and so React does not rerender.
Our code change does result in a rerender, but because it causes
left/rightActive to update.
I guess this works fine as long as it's ok to deactivate the sidebars
when transitioning to/from a small viewport, but idealistically
we should put this.isSmallViewport in this.state.