Fix obscure CustomSelect production bug on initial load;
On **initial** page load in **production** (gatsby build && gatsby serve),
CustomSelect does not show the green v checkmark of the selected option.
This works fine in development, or in production once another option is clicked.
This seems to be caused by the exported sass variables which are undefined
on the initial page load; the solution is to assign them in componentDidMount
instead of in the constructor.

This is also what caused the bug in 7d12c8061707c4eae1adacdfb3f4488339eb7450
(7 commits back).