Fix FOUC in production;
This was caused by setting a theming class on body in componentDidMount.
By removing this class and setting it on #layout in the render() method,
everything follows the more conventional pattern, and this problem does not occur.

We are however still left with a 'Flash-Of-Unstyled-Fontsize' due to
the .noTouchscreen class changing the font-size when componentDidMount
adds it to the body.