Remove React.cloneElement in favour of React Contexts;
This way, downstream users can move Leftside, Rightside and Mainview into
custom components in other files. This was not previously possible due to
the displayName check, which would NOT be set if <Layout>'s children are
custom components (that render a theme component). The result of this was
a completely blank page when using custom components ...