CSS modules generally not worth it for theme stylesheets;
The complexity overhead is huge due to the fact that the class/id names need to be passed
to all children components, which would not be the case for normal component stylesheets.
We will rewind the CSS module implementation because the additional complexity is not worth
the single scope benefit. The code will be much more readable by just creating longer
classnames. Note that the imported mediumViewport sass variable will need to be removed,
because ':export' is not technically part of the sass spec, but of the CSS modules spec (ICSS).